refactor: document and namespace the loader file
